requirements.txt
streamlit
opencv-python
matplotlib
pillow
moviepy
scipy
tensorflow
fer
